refactor(frontend): split services by feature
